Web App/requirements.txt
tensorflow-cpu==1.15.0
flask
opencv-python-headless==4.2.0.32
numpy
werkzeug
pillow
uuid
urllib3